requirements.txt
pyfunctional>=1.5.0
langcodes>=3.3.0
language_data>=1.2.0
